Add support for generic payloads - Add `MAV_CMD_GENERIC_PAYLOAD_FUNCTIONS_CONTROL` and `GENERIC_PAYLOAD_STATUS` AND `GENERIC_PAYLOAD_ERROR_FLAGS` to development.xml
`MAV_CMD_GENERIC_PAYLOAD_FUNCTIONS_CONTROL` is used to control payload functions. It can be used for simple control such as ON/OFF of a flashlight. It can also be used to control multiple functions of a component such as arm/disarm and deployment of a parachute.

Topic: generic_payload